app/requirements.txt
dash-bootstrap-components==1.5.0
dash-mantine-components==0.12.1
pandas==2.2.1
dash_ag_grid==31.0.1
dash-renderjson==0.0.1
jsonschema==4.21.1
girder-client==3.2.2
tqdm==4.66.2
jsonschema==4.21.1